장바구니 구현

+++ RaR/src/main/webapp/js/cart.js
@@ -1,4 +1,5 @@
 $(function(){
+	
 	//목록 호출
 	function selectList(){
 		$.ajax({
@@ -8,30 +9,33 @@ $(function(){
 			success:function(param){
 				$('#output').empty();
 					let head_output = '<tr>';
-					head_output += '<th><input type="checkbox" id="selectAll"></th>';
+					head_output += '<th><input type="checkbox" id="selectAll" checked></th>';
 					head_output += '<th>상품정보</th>';
 					head_output += '<th>가격</th>';
 					head_output += '<th>상태</th>';
 					head_output += '<th>삭제</th>';
 					head_output += '</tr>';
 					$('#output').append(head_output);
+				
 				$(param.list).each(function(index,item){
 					let output = '<tr>';
-					output += '<td><input type="checkbox" data-cartnum="' + item.cart_num + '" class="selectCheck"></td>';
-					output += '<td><img src="' + contextPath + '/upload/' + item.bookVO.bk_img + '" width="80">' + item.bookVO.bk_name + '</td>';
+					output += '<td><input type="checkbox" data-cartnum="' + item.cart_num + '" class="selectCheck" checked></td>';
+					output += '<td><img src="' + item.bookVO.bk_img + '" width="60">' + item.bookVO.bk_name + '</td>';
 					output += '<td> 정가 : ' + item.bookVO.bk_price
-							+ '<br> 판매가 : ' + item.itemVO.item_price + '</td>';
+							+ '원<br> 판매가 : <span id="item_price_' + item.cart_num + '">' + item.itemVO.item_price + '</span>원</td>';
 					output += '<td>' + item.itemVO.item_grade + '</td>';
 					output += '<td><input type="button" data-cartnum="' + item.cart_num + '" value="삭제" class="delete-btn"></td>';
 					output += '</tr>';
 					
 					$('#output').append(output);
 				});
+				cartTotal();
 			},
 			error:function(){
 				alert('네트워크 오류 발생');
 			}
 		});
+		
 	}
 	
 	//장바구니 삭제 함수
@@ -56,13 +60,14 @@ $(function(){
 			}
 		});
 	}
-		//전체 선택
+	//전체 선택
 	$(document).on('change', '#selectAll', function() {
 	    if ($(this).prop('checked')) {
 	        $('.selectCheck').prop('checked', true);
 	    } else {
 	        $('.selectCheck').prop('checked', false);
 	    }
+	    cartTotal();
 	});
 	
 	//장바구니 개별 삭제
@@ -73,11 +78,61 @@ $(function(){
 	});
 	//장바구니 선택 삭제
 	$(document).on('click', '#selectAll-btn', function () {
-	    $('.selectCheck:checked').each(function () {
-	        let cart_num = $(this).attr('data-cartnum');
-	        deleteCartItem(cart_num);
-	    });
+		if(confirm('선택한 상품을 모두 삭제하시겠습니까?')){
+		    $('.selectCheck:checked').each(function () {
+		        let cart_num = $(this).attr('data-cartnum');
+		        deleteCartItem(cart_num);
+		    });			
+		}
 	});
+	//선택한 상품 총 구매비용 계산
+	$(document).on('change', '.selectCheck',function() {
+		cartTotal();
+	});
+	//구매비용 계산 함수
+	function cartTotal() {
+	    let totalCount = 0;
+	    let totalPrice = 0;
+	    let ship = 0; // 배송비
+	    let totalPayment = 0;
+	    let totalPoints = 0;
 	
+	    // 각 상품별 정보를 가져와서 계산
+	    $('.selectCheck:checked').each(function() {
+	        let cartNum = $(this).attr('data-cartnum');
+	        let item_price = parseInt($('#item_price_' + cartNum).text());
+			
+			//상품 개수
+			totalCount++;
+			//상품 금액
+	        totalPrice += item_price;
+			
+	    });
+		
+		//배송비	
+        if (totalPrice < 30000) {
+            ship = 4000;
+        }
+		//포인트
+        totalPoints = Math.floor(totalPrice * 0.01); // 총 주문 금액의 1%를 적립
+		//총 상품 금액
+	    totalPayment = totalPrice + ship;
+	
+	    // 결과를 화면에 업데이트
+	    $('#cart_total td:eq(0)').text(totalCount + '개');
+	    $('#cart_total td:eq(1)').text(totalPrice.toLocaleString() + '원');
+	    $('#cart_total td:eq(2)').text(ship.toLocaleString() + '원');
+	    $('#cart_total td:eq(3)').text(totalPayment.toLocaleString() + '원');
+	    $('#cart_total td:eq(4)').text(totalPoints.toLocaleString() + '점');
+	    
+	    //전송할 데이터 값 입력
+	    $('#totalCount').val(totalCount);
+	    $('#totalPrice').val(totalPrice);
+	    $('#ship').val(ship);
+	    $('#totalPayment').val(totalPayment);
+	    $('#totalPoints').val(totalPoints);
+	}
+	
+	//초기 함수 호출
 	selectList();
 });
